Changed the constructor for GameStat class, renamed GameStatStructTests --> GameStatClassTests &removed some tests in the GameStatClassTests.
